feat: Add how to for elo evaluations refactor: Move elo evaluation logic from "elo_evaluator.py" into "incremental_evaluator.py"
TASK: IL-502
