+++ src/intelligence_layer/examples/qa/elo_qa_evaluation_logic.py
@@ -7,11 +7,10 @@ from liquid import Template
 from intelligence_layer.core.detect_language import Language
 from intelligence_layer.core.model import CompleteInput, CompleteOutput, ControlModel
 from intelligence_layer.core.tracer.tracer import NoOpTracer, TaskSpan, Tracer
+from intelligence_layer.evaluation import EloEvaluationLogic, MatchOutcome
 from intelligence_layer.evaluation.dataset.domain import Example
-from intelligence_layer.evaluation.evaluation.evaluator.elo_evaluator import (
-    EloEvaluationLogic,
+from intelligence_layer.evaluation.evaluation.evaluator.incremental_evaluator import (
     EloGradingInput,
-    MatchOutcome,
 )
 from intelligence_layer.evaluation.run.domain import SuccessfulExampleOutput
 from intelligence_layer.examples.qa.single_chunk_qa import (
